Update README.md
This repository contains "RuPersWordAssociation", a dataset of  association reactions to 50 Russian words most important for Russian  linguistic culture. 
RuPersWordAssociation dataset  is, to the best of our knowledge, the largest existing (at least from publicly available) word association database with plenty of metadata (psychodemographical information about the respondents, reaction time, semantic similarity metrics) and linguistic annotation for the type of a relation between a cue word and reactions (more than 22000 word pairs were annotated).  
A detailed description of the dataset and the results of linguistic annotation and present the outcomes of the experiments which revealed individual preferences for certain types of associations are presented in the manuscript "RuPersWordAssociation: a new dataset to study individual association behavior' (Litvinova et al. 2024) submitted for review for IMS 2024.